+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Renamed `master` branch to `main` - Fixes [Issue #393](https://github.com/PlagueHO/CosmosDB/issues/393).
 
+### Added
+
+- Added support for specifying custom endpoint in `New-CosmosDbContext` to
+  support alternative clouds - Fixes [Issue #395](https://github.com/PlagueHO/CosmosDB/issues/395).
+
 ## [4.2.1] - 2020-06-15
 
 ### Fixed